Cleaning up Code
