+++ src/App.js
@@ -1,13 +1,20 @@
-import React, { useState } from 'react'
-import { Route, Switch, useHistory } from 'react-router-dom'
+import React from 'react'
+import { Route, Switch } from 'react-router-dom'
 import Form from './components/Form'
 import Dashboard from './pages/Dashboard'
 import useReviews from './hooks/useReviews'
 
 export default function App() {
-  const { reviews, addReview, removeReview, editReview } = useReviews()
-  const [editReviewState, setEditReviewState] = useState(undefined)
-  const history = useHistory()
+  const {
+    reviews,
+    editReviewState,
+    setEditReviewState,
+    addReview,
+    removeReview,
+    editReview,
+    onEdit,
+  } = useReviews()
+
   return (
     <Switch>
       <Route exact path="/">
@@ -23,11 +30,4 @@ export default function App() {
       </Route>
     </Switch>
   )
-
-  function onEdit(id) {
-    const editedReview = reviews.find((review) => review.id === id)
-    setEditReviewState(editedReview)
-    editReview(editedReview)
-    history.push('/create')
-  }
 }
